pc/requirements.txt
# Novita.ai GPU Instance Manager - Python Dependencies
# Phase 1 + Phase 2 Requirements

# GUI Framework
PySide6>=6.4.0

# HTTP Client for API requests
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# Logging (built-in, but listed for clarity)
# logging - standard library

